pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "paperless-firefly"
version = "0.2.0"
description = "Extract finance data from Paperless-ngx and import to Firefly III"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "requests>=2.28.0",
    "pyyaml>=6.0",
    "django>=4.2",
    "lxml>=4.9.0",  # XML parsing for e-invoices
    "httpx>=0.25.0",  # Async HTTP client for LLM API
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "responses>=0.23.0",
    "black>=23.0",
    "mypy>=1.0",
    "types-requests>=2.28.0",
    "types-PyYAML>=6.0",
    "types-lxml>=2023.3.28",
    "reportlab>=4.0",  # PDF generation for test fixtures
    "isort>=5.12.0",
    "ruff>=0.1.0",
]

[project.scripts]
paperless-firefly = "paperless_firefly.runner.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"paperless_firefly.review.web" = ["templates/**/*.html", "static/**/*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100
known_first_party = ["paperless_firefly"]
skip = ["__pycache__", ".git", ".venv", "venv"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011", "B017"]  # Allow assert and blind exception tests

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
